requirements.txt
qbittorrent-api
rich~=14.1.0

requests~=2.32.4
urllib3~=2.5.0

rumps

python-telegram-bot

PyQt5

pyobjc-framework-Quartz
pyobjc
PyOpenGL
PyOpenGL_accelerate